ENH: re #4, the labels imported are now associated with the volume
I think it is appropriate to allow only one volume to be present in the
annotation at a time, and to have all markup elements associated with the
volume to refer to the same finding in that volume. This is consistent with
the capabilities of ClearCanvas AIM workstation, and seems to be in line with
the intended purpose of AIM 3.0. Simplifies our task too.