BUG: Fix and test casting of resampled label map.
Change to a vtkImageCast to simplify the setting of the resampled label type.
Cleaned up the logic, removed an extraneous AddNode (the scalar volume reader
adds the input node to the scene).
Removed debugging print outs.
Removed the try block as it was masking errors.
Added two tests for the converter script, with new label map volumes that were
created from the test DICOM series. The second one has a rotation applied to it
to trigger the reformatting.
Use sys.exit to ensure that tests will be detected as failing.

Issue #42