BUG: ensure label volume is unsigned short, add -F cmd line option
From discussions on the issue, it's a requirement that the label volume be
unsigned short, so check and cast it before checking for geometry mismatch
(the geometry mismatch check does not catch image data scalar type mismatches).

Add a command line option to the converter script, -F, so that if there is a geometry
mismatch, the script will automatically do a resample. If there is a geometry mismatch
without this argument present, abort the script.
Updated the test to use this flag so that it will pass.

Issue #42