BUG: update test to use new label map class
The test was getting all the scalar volumes and checking for the
label attribute, but the change to use a new class, vtkMRMLLabelMapVolumeNode,
meant that the sought after label file was not being found to check.
This change updates the test to get just the label maps in the scene and
to test that there is one and it is correct.