ENH: add color table and terminology for DICOM segmentation reading
When reading in a DICOM segmentatin file, create one color table node
for all the segments, and build up a terminology associated with the
segmentation from the .info files.
Adjusted the logic call to not use the color node, do the terminology and color
information file parsing in the segmentation plugin.
Added a check that the SEG2NRRD CLI was loaded to prevent a crash if Slicer was started up
without CLI modules.

Issue #52