ENH: accept RGB display value in place of CIELab
RGB is arguably more common color scheme than CIELab, and the one used in Slicer.
Conversion functions used are the same as in SEG2NRRD (this also makes the process
symmetric, since SEG2NRRD outputs RGB colors).

In the future, it is suggested that the conversion routines move somewhere else
to be more accessible, perhaps DCMTK is the best place for this.

This should resolve #62.