ENH: update to use new terminology methods, order
Update the plugin to use the new order of terms when adding a term to
a terminology.
Update the SEG exporter self test to use the new color logic methods,

Depends on Nov 12/15 update to:
https://github.com/naucoin/Slicer/tree/4047-Add-semantic-color-information-to-color-selector-in-editor

Unify terminology file formats:
Update SEG2NRRD.cxx to use the format from EncodeSEG.cxx.
Update DICOMSegmentationPlugin.py to parse the single line
format, with the capability to deal with multiple lines defining
terminology for different labels in one file.

Issue #52