BUG: load all slices of the referenced series
An issue that was not considered is that the segmentation frames may not be
contiguous, and thus the referenced instances may not result in the contiguous
slices for the scalar volume. TODO - this should be fixed in the Slicer core -
probably the referenced scalar volume should not be loaded if there are warnings
generated by examine().

For now, to resolve this issue, all of the instances of the referenced series
are added to the loadable, and the whole volume of the referenced series is
supposed to be loaded.