BUG: copy DICOM files to a temp directory for SEG conversion
There was a cryptic runtime error in Slicer on Windows, and I thought the most
likely reason is the exceedingly long command line (the length for a whole body
CT segmentation was way way over the Windows limit of ~8k). Copying the files
into a temporary directory and using the directory parameter resolved the
issue. Temporary directory is removed after conversion is done. In the future,
we should address this by allowing to pass list of files in the JSON directly
in dcmqi (issue created https://github.com/QIICR/dcmqi/issues/297).