Do not set default terminology in loaded segments
While loading a DICOM segmentation object, if category or anatomic region was not defined then the DICOMSegmentationPlugin added
a default ['85756007', 'Tissue', 'SCT'] category+type and ['38266002', 'SCT', 'Entire Body'] anatomic region.
However, this resulted in importing+exporting a segment changing the content. It also produced a warning (because the added terms
were not found in the segmentation's context).

This commit changes the behavior so that if the category, type, or anatomic region in the original segment are undefined then they
remain undefined in the loaded segmentation.