Set NameAutoGenerated property for loaded segments
NameAutoGenerated of a segment means that the name was automatically generated from the terminology code meaning.
This helps in preserving names that the user manually types for the segment.

Previously, when a segmentation was loaded the NameAutoGenerated flag was left at default (true), therefore when the
user modified the terminology then the name got automatically overwritten.

Fixed the issue by setting NameAutoGenerated to true only if the name is generated automatically from the
terminology.